requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
cloudinary==1.36.0
crispy-bootstrap5==0.7
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.7
django-autocomplete-light==3.9.7
django-ckeditor==6.7.0
django-cloudinary-storage==0.3.0
django-crispy-forms==2.0
django-dbml==0.6.0
django-debug-toolbar==4.2.0
django-js-asset==2.1.0
django-taggit==4.0.0
Faker==19.3.0
fonttools==4.43.1
fpdf2==2.7.6
gunicorn==21.2.0
idna==3.4
jsmin==3.0.1
packaging==23.2
Pillow==10.0.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6
whitenoise==6.6.0
